Warning compilation
Former-commit-id: 61ae00e7597ca0fcde98b10f22cb176868d45e00

+++ libraries/parallel/mpi_angular_class_average.cpp
@@ -263,7 +263,7 @@ void MpiProgAngularClassAverage::run()
                 MPI_Recv(lockWeightIndexes, lockWeightIndexesSize, MPI_DOUBLE, MPI_ANY_SOURCE, TAG_I_FINISH_WRITTING,
                          MPI_COMM_WORLD, &status);
 
-                lockIndex = lockWeightIndexes[index_lockIndex];
+                lockIndex = (int)lockWeightIndexes[index_lockIndex];
                 //#define DEBUG_MPI
 #ifdef DEBUG_MPI
